New logo
The application of the logo is still in progress, but the log itself is currently looking solid. Added a possible favicon and a logo to the website's navbar. Changed the background color of the site to something more neutral (still not married to this color; just temporary)  and fixed a broken link to the main page. Also discovered some bugs on mobile and updates TODO comments to reflect them.