Fixed mobile section heights
They are now more dynamic and grow as necessary. Only problem with the current method is when I tested viewing the site in desktop mode on Chrome mobile it was very ugly because nothing scaled properly, but I can work on that later. Next step is to get the mobile menu functioning properly.