Progress towards profile page
I made this branch because I don't know if I am going to continue to use Firebase due to the unnecessary difficulties in user management and tracking. Right now I have the foundations of a profile page (a PHP file just in case; might change that) but suddenly I can no longer access the user's data at all. Saving this primarily as a backup so I can try other avenues.