Email login and view user data
This certainly isn't everything I had in mind, but it is proof that Firebase is capable of working if you know what you're doing. This upload is more of a progress point than anything just because I'm about to go to the gym.

Anyways, the next step is to turn this into a working profile page and to see what I can do with it. Also adding social media login if possible with this method. One thing to note is you cannot sign in with email if there is an existing social account in the database with your email address using this approach.

What I really need to do is distill the method of user tracking and see if I can apply it to the main branch.