Set up CI with Azure Pipelines
I'm hoping this means I can easily deploy my code to Azure. The main problem is that I will now have to make commits to test my code, but hopefully I can still set up something with FTP and just have this redeploy the code officially with new GitHub releases. I'm hoping this doesn't take three minutes to run again, because if it has to reinstall Node every time or something that is going to be a problem. [skip ci]