Track user auth state
Even though it only works with email so far, this is definitive proof that the user auth state can be tracked without much difficulty in Firebase. The next step is to combine this with the social log in options and the profile page, which would make for a great merger with the profile and master branches.