Transition to PHP fragments
If nothing else, this will simplify the development process by allowing me to make instantaneous changes across the entire UI. I also transitioned the app from desktop and mobile to entirely mobile-centric so there will be no screen size issues later down the line. Several changes will not make sense right now, such as the removal of scrolling on menu click, but they are just a step in the mobile direction for later down the line.

I may need to replace PHP with jQuery mobile or something similar when moving to a mobile framework because it might not be supported. At the very worst there should be some automatic method to replace the PHP include statements with the code on deploy. Several possible fixes to this were included in the header comments of index.php