User profile pictures displayed
Now the nav functions like it should and switches to displaying the user profile picture once the user signs in. The user is also assigned a default profile picture on account creation if their platform of choice doesn't come with one. 

Next steps are to get delete account and sign out buttons working properly on a profile page.