Attempted speed-up of data retrieval
It *most likely* won't be an issue in production since the majority of the scripts will be downloaded with the app, but right now the loading time is huge and an insurmountable issue for the profile page. These changes sped up load time considerably (~5 seconds to ~2) but I don't think that's going to be enough.