Prototype demonstrating Auth0 speed
It is much faster and hopefully easier to work with than Firebase. This demo doesn't have all the old styles, but so far it has functioning log in, log out, and a profile page, which is more than I could say for Firebase because the profile page was giving me such a hard time. Going to continue working on this, even though it won't be free in development after a month, because I really think this is the way to go.

Expect a lot of changes and restructuring in the near future, as I still need to finish converting from Firebase and all my old files (which I safely backed up before this commit). I am taking my time because it is hard to know off the bat which elements of the example code are necessary and which are optional.