Getting nav bar look back
It wasn't easy, but I managed to find the source of the problems with breakpoints for the navbar-toggler and remove them all so the profile picture displays the entire time. Currently just commenting out most of the things that I don't need in auth0-theme and I'll remove them later if it doesn't cause any problems.

The next step is to continue integrating some styles until they are at a point I think will be easier to work with in the future. Also getting rid of more HTML from the example that I don't need. After that, I want to start seeing if I can get a mobile framework to work for me so I can stop stressing about nothing being able to convert to an app.