UI beginning to display properly
When I started there was a plethora of issues with files not being where they were supposed to be once deployed to Android and tons of failed retrievals of online scripts, but I managed to fix the majority of that now. The login is currently not functioning and the UI looks shitty but I'm about to work on all of that.

There are several things happening that are temporary, such as the Phonegap logo displaying on startup while the device isn't ready instead of my Coupon Booked splashscreen I need to design, but I'm not going to worry about any of that until I can get this login working the way it should.