Barebones mobile login
I scrapped my UI entirely (hopefully for the time being) because this proves that the login can work as expected on mobile. My next step is to merge this into the main branch, as it has served its purpose in proving that Cordova is capable of supporting this app, then to get the UI back into shape and continue with development.

To be completely honest, I have no idea if webpack is even used at all in this app, as I don't think I've needed it yet, but I guess we'll just have to see what happens. I'm assuming Babel has something to do with Webpack, but I did not add the Babel config or dev dependencies, so Webpack might not even work if I do end up needing it.