Profile dropdown redirect
There is still no use for the dashboard button, but I am leaving it for now just in case. I have not yet tested the app on different screen sizes, so I do not know if the removed code was of any importance.

Next step is to get the profile page UI up to par with that of the home page and to continue fixing the style of the nav bar and dropdown menu.