Removed scrolling over navbar
The page element was positioned absolutely in app.css when it shouldn't have been. I am using an extremely hacky method of getting around the problems of a fixed navbar right now, but I have a lot of functionality to implement and I will cross that bridge when I get to it.