Prevented 429 error
The problem was that the request for the same data, such as the profile picture for the nav bar, was being made every single time the route changed. I solved it by storing the data in a variable, which *may* have side effects down the road, but it serves its purpose right now. Also removed useless CSS libraries.