Implement refresh tokens
Previously, the app would kick the user out when the JWT token expired after 10 hours, but now (in theory) the user can remain logged in indefinitely. The nav code now includes a check to see if the user is not authenticated and sends them to the login page if they are not, even though the refresh tokens should have resolved that issue. I browsed through the possible Cordova hooks and learned that the only hooks I would ever need are Aut0 hooks, so I removed the Cordova hook documentation.

Also implemented a method of getting the user ID, so my next step is to use that to create unique users in my SQL database for storing user coupon books.