Resolved issue with login redirect
While the retrieval of an access token after the previous one expires works perfectly, the function it is in is currently not asynchronous. This led to loading a new access token working sometimes and sometimes receiving the code too late.

When that happened, I noticed that the fail-safe "redirect to login" was not working as planned because I had attempted to access the nonexistent "this" object. Now I resolved that and I plan on making the token retrieval asynchronous to avoid any issues in the future.