Create CB UI initialized
Now the front-end framework necessary to begin interacting with the back end is in play. The next step is to style these things in an attractive manner, which shouldn't be difficult since I can borrow most of it from Jessica Wants, then to actually begin interacting with the back end.