Successful database interaction
Alright, this commit is a big one and a little disorganized, so bear with me here.

First of all, I used PHP to interact with the MySQL database. The PHP files are stored in the website directory because they have to be on the web server, and I'll interact with them via AJAX from Node. 

Secondly, I removed the directory password protection from htaccess because the PHP files need to be accessible to send queries. 

I have not yet finished setting up all the necessary interactions with the database. So far I have only gotten getTemplate working the way I want, but the others are not far behind. The next stage aside from that is to get the sliding menu UI set up for the dashboard, which will rely on the DB queries. After that comes the ability to create and modify coupon books, which will be a bit of a doozie.