All PHP files doing their job
Pulling user data, modifying data, and creating new coupon books all behaves exactly as planned. They do not yet interact with the front end and there is no real data generation to work with yet, but the majority of everything is solid on the back end.

The app file's functions were all updated to reflect the PHP files, and some errors were discovered as well that need to be fixed.