Dashboard tab menu
The menu does everything I wanted it to do except being swipeable, but now that I've gotten it this far it shouldn't be too difficult (fingers crossed) to use a library to implement swipe detection.

To get the tab menu how I wanted it, I simply had to gut the 9000-line Materialize CSS file down to 110 lines. Piece of cake. Next on the docket is to look into minimizing the file size, both mine and Node modules, because it is overly bloated currently.