Gutted minimize.js
I went through and got rid of all the functions that didn't apply to what I was using in this project, reducing the file from 12375 lines to 5723. I found the functions relating to the tab menu, so hopefully in the future I will be able to modify them and add an animation that works in Cordova. The native one for Materialize was too bloated and slow.

In main.css, the z-index inclusion fixed elements scrolling above the menu. Also moved all the swipe functions from the route to a separate function to make work easier.