Resolved login expiration issue
Now, FINALLY, the login problem *should* be gone. The issue was that even with the Promise being handled, the render code was being called before the successfulAuth function could complete. I simply had to move the render call into the success and failure functions and it works fine.

The only issue I see that may be a problem is it appears the expiration code may be running when the access token is not expired, but if I continue to notice this behavior in the future I will address it then.