In-app notifications to replace alerts
Pretty self-explanatory. Not the world's prettiest solution, but it gets the job done. Can tweak it more later, such as coloring, animations and positioning. Right now my CSS is overriding the default, so I might just go in and fix the default for my purposes since it isn't that long. Will likely use the unminimized JS file as well to modify animations among several other things.

Also modified the PHP for createBook to properly fail. The only situation I set up for is if the POST variables don't exist, which should never happen, but I'm ready in case it does. Learned how to fail an AJAX request properly, so I'll start implementing that a lot more in the near future.