Notifications display properly over content
I had only tested them on the empty manipulate page, so I didn't know how they reacted to actual content. I ended up having to absolutely position them so they display properly over everything. I also made the shift to the unminified CSS so I could do things the right way without workarounds.

I did remove all the different positions from the CSS aside from the default since I'm manipulating that to be in the center anyways, but that shouldn't be a problem and if it is I can always redownload the library. Also added a notification that I forgot about in pullUserRelatedBooks.