Create and update templates
I know I already have a commit for creating templates, but this is an overhaul of that because I rethought the process. What happens now is a blank template by the specified name is created with a specific empty data schema and you will populate it from the graphical interface. On the click of the plus button, you will be redirected to a form that you fill out with the coupon information. After submitted, createCoupon will take that data and push it to the local book variable, which will be uploaded to the server when the save button is clicked.

updateTemplate does exactly what it sounds like. It is what is called when you click the save button when in development mode, vs. updating a normal book when you aren't via updateBook. Next plans are to get that form working and look into overhauling the create page with template cover pictures, so we'll see how that goes.