Form data retrieval for coupon creation
Included a decent-looking form to start, not the greatest but it'll do until I get around to things like not letting the page move when the keyboard is opened. 

So far there is no data checking whatsoever and changing and retrieving the image is not supported, but this is a definite step in the right direction.