Notifications slide in from bottom
Eventually I want to optimize the SimpleNotification library from the ground up with support for notifications at the top center and bottom center, but until then all these minor workarounds are going to have to do. Need to eventually rewrite it to improve it though because anybody else who looks at this part of the project will get REAL confused. If that's you, hey :) sorry.