Start of sending implementation
I decided on the method by which coupons should be sent, and this is the beginning. Obviously, as seen by the major TODO, there is still work to be done, and decisions to be thought out, but this is a major step in the right direction.

Also implemented GET and POST variable checking across all other PHP files to knock one more thing off the ever-growing TODO list.