Share code UI and routing
Everything about the share code now works as it should except for it actually doing anything. You can copy the share code by clicking on it and there is a pretty tooltip letting you know it is copied.

What is left for sharing is adding native share support to the orange button, small UI tweaks way down the line, and a way to get to the share screen easily from a book.

Also made the errors for failed Ajax display useful information rather than just the error code. The exit message from the PHP is now visible in the console.