Edit and add coupons
Not only have 98% of the UI kinks been worked out, but everything functions as it should. Cleaned up some of the pullUserRelatedData code as well because it was giving me a serious headache.

Next on the todo list is the ability to delete coupons and a redeem code route. Hardest thing i need to do soon is the new image UI, which I will probably put off because it will be a very involved process.