Redeem share codes
Got a pretty decent UI and route for it, as well as a working backend that performs all the necessary checks. Only other thing I need to handle on this front is displaying the receiver in a more attractive way, such as by fetching the ID holder's name if possible. Might have to store that in a separate SQL column if nothing else.

Another thing I want to do is hide the save button on the view coupon page because it serves no purpose and is misleading, but hopefully that won't be too difficult.