Decompose and clarify displayBook()
The code had gotten much too long and repetitive, so I broke it down and made it readable. There is very little repetition now and I was even able to fix a few pesky bugs I didn't even know about. I also moved the coupon functions close to all the functions where they are displayed because that method of grouping made much more sense and saved me a lot of scrolling and searching.