Delete books and miniPreview book UI
Books can now be deleted with the trash can icon in the new mini preview that displays book information once the book is selected from the dashboard. Deleting books and coupons now displays a prompt that requires user confirmation.

When a user attempts to save a book without changing anything, they will be notified and it will not be saved. Attempting to leave the page with the back button when the book has not yet been saved will require user confirmation, and this feature will soon be implemented for coupons as well.

Users now cannot create coupons with non-unique names and cannot edit the names to be the same as an existing coupon.