Natively share code
Implemented the native share plugin and it works as promised. Eventually might need to fine-tune the message and everything, but the important part is that it works. Also replaced the iframe Auth0 button with my own replication, which wasn't as hard as I thought it would be when I went through all that trouble before.