Redeem coupons and update server
Redeeming coupons in a book that was sent to you works as it should and decrements the count locally and on the server, but does not yet send a notification to the sender. That is now the top priority, as is making sure the uncommitted login code works because if a user doesn't stay logged in then they won't get the notifications.

I did not test several things like the count being zero, but I'm sure they work as they should and if not I can test them later.