Protected unauthorized template access
Made it so my user ID, set in env.php, is the only valid one to modify template data. Need to keep my ID secure now, if that's possible. Might even be able to use hardware identification in the future so only my phone could update data. That would be harder to determine and spoof.