Create button moved to dashboard
I'm beginning to revamp the home page and the first step was removing that incredibly ugly button that I was never going to end up styling. Now it's a lot more intuitive to create books, and navigation is easier because the back button is supported on the create page and the dashboard. Also added a neat little shadow to the button for dramatic effect ;)