Speed up deployment
This will likely have to be reinstated before official release, because even though -p was problematic with --optimize-minimize it vastly reduces the file size for user download. However, in the meantime this is epic because you can actually see what lines the console output is generated from because it isn't minified.