Edit book details
Obviously image editing hasn't been done yet, but the rest of it works nearly as perfectly as can be expected. The animation when opening the goo menu is sometimes choppy, but I'll have to work on resolving that later, possibly with something like https://codepen.io/ajmeese7/pen/abzZLxK. Next step is to find an attractive way to show the book description, probably with like two lines and ellipses to open a popup that shows the rest as well as all the book details. Could be super attractive like the Stripe form.