Edit book details and image
Added the ability to choose images from device and take pictures directly in the app. Both must be cropped to a square after being selected by the user, but it might be possible to make the two more uniform in the future. The image features were just added to the book edit page so far but my next move is to implement it for coupons as well. I'll also be adding the uploading of images to the Cloudinary server, I just need to think more on which stage to do that at. Shouldn't be tough.

The whole project was fucked up when I tried to use a different library for images so I had to rebuild the dev environment from scratch. Had to make some funky changes to get it to work with the latest version of Android Cordova, but it all worked out. Lesson learned: always read the issue page BEFORE installing a sketchy library. This commit is more of a backup bridge in case I fuck things up again, but I think I can power through an get it all finished up soon.

NOTE: Not sure about removing the minimum SDK version of Android in config.xml. Going to look further into that soon.