Image upload working
I can now upload images automatically after selection and cropping. Will be super easy to implement for taking pictures as well then for coupons. I went fancy with it and organized the images on the server according to userId and book v. coupon, but I'm going to think more on it if there's a more specific structure I want. Like if I want to tag images with certain metadata to make it easier to delete by category on user deletion and stuff like that.

It's a shame nobody will ever see how nice this part of the app is. Putting all this work in now so hopefully I can make my life a lot easier down the line once I (hopefully) implement bug reporting but we'll see.