Tempates pulled from server
Believe it or not, that was the easy part of the commit. All the hard stuff was reorganizing the code so the debugging flow made sense with related functions near one another and fixing quirky UI bugs because certain functions were called in the wrong order. That's why I need to implement testing or use some alternative method from a Medium article I just read. This, of course, replace the getTemplate function, rendering it obsolete.

This commit uploads template images (in development mode) to a separate folder location on Cloudinary, and with only a small bug that shouldn't be hard to rectify soon. Whenever user books are being manipulated their images are uploaded to the template folders too, which should never even happen but developer users can be stupid too (guilty as charged ;)). 

I had to renovate the limitDescriptionLength function so it would function with new coupons and old ones. Kinda hard to explain but look at the differences and you'll maybe get it a little more.