Switch to template literals
Also fixed an issue with notificationOptions not being in the global scope, and env not being available to sent books. Left some of the strings as they were because template literals looked ugly with them, and I probably got overzealous and switched some stuff that didn't need to be transferred over, but I'll worry about it later.