New coupon image upload
That temporary fix a few commits ago was replaced with a permanent solution that's a lot more intuitive for the user. Made a few function names easier to follow and removed some useless old comments as well.

Also got rid of all the functions commented out in exporting because that was a dumb idea. Now it just has what you actually need to export.