Template image upload folder correct
Now it manages to distinguish between when an idiot developer uploads an image to a user's book in development mode or the actual templates.

For some reason the development variable was undefined outside the helper file so I had to move it to sentBooks for now. No real loss since that's the only place it is used.