Slight PNG upload improvement
Now it doesn't even show the cropping screen with PNG files since it's the uncropped image that will be uploaded anyways at the time being. So until I can find a way to incorporate that other plugin that allows rotation and transparency, I'm just not going to let them crop PNGs at all.