Custom home button listeners
Now whenever you press the home button on the book page or coupon page it checks to see if changes were made and prompts you about whether you want to discard those changes to leave the page like if you pressed the back button normally. The next step is to implement something similar for the navigation dropdown, but I haven't entirely decided if that's the route I want to take. Also won't really be a problem if I implement the autosave method I was thinking about, so I'll just wait instead of committing twice.