More intuitive buttons
Now a create button is displayed instead of a save button on initial retrieval of a template, and a lot of similar stuff like that. Removed some old print statements and added proper handling of development mode vs normal throughout.

Only thing do add relevant to this commit now is a confirm discarding changes interface for leaving a template after modifying it, in normal mode because dev uploads changes immediately to the server.

Also going to add in template creation eventually but that's hardly ever going to be used again so release is the top priority as well as generating content.