Fixed jQuery issues
I had input the parameters wrong on one of the functions so it was getting all huffy and puffy every time I would try to load a page in. Also made some small issues with image uploading go away, and may have discovered some others that require fixing. To be honest I did this a few days ago and forgot the purpose so I'll get back to you when I discover if that comment still holds any relevance.