Start of website revamp
Removed the ugly-ass useless mobile menu I spent so much time on, because they can just scroll through it if they want to look at it. I added a Google Play Store button that doesn't link to my app yet since it isn't on the store but it can be easily modified in the future.

I started updating the example images with real images but I'm not sure what to write yet or if I'm actually a fan of it. On mobile I should probably also implement a 'display more' option next and just show two at first or something since that's so much scrolling. Also updated the 404 page with a link to my GitHub to contact me with any issues that may arise.