Startup animation
I was able to modify the Codepen to show when the app is loading up, and formatted the text so it is legible. Not the most perfect thing in the world, but it's pretty enough. Can modify easily in the future if I see fit. Also added disabling the animation to the settings page since it was looking mighty empty.

Next step with the animation is to cut out the separate route entirely so I avoid that split second of white screen lag and just show it all at once.