Loading icons
Instead of staring at a blank screen waiting for your content to load in, now there are loading icons that show for a set time before disappearing to hopefully reveal your content already loaded in. It really depends on your signal strength and the amount of content you have to load in, but I'd say it's a fair guestimate.