Stripe button finished
Instead of having that strange hacky solution, now this one sort of makes sense. I almost used the Stripe code the way it was intended, which is a step in the right direction. I removed some of the styles applied by the library to give it a bland appearance that blends in with the rest of the map.

Also added back button support for closing out the popup instead of going back a route and still having the overlay visible. Next step is to follow up with all the added TODO messages.