Native in-app purchases
I read the terms and conditions of the Play Store and they are super against me using Stripe, so I started using a plugin that handles most of the heavy lifting associated with that for me. I have no way of knowing if it actually works until I can get it released for alpha testing and actually try it out, so fingers crossed.

I did a massive refactor intuitively with next to no testing for functions associated with the sharing of books, so hopefully I didn't fuck anything up too badly. I also finally started VERSIONING in the package.json!!! Crazy, right?

I'm pretty bummed that I did all that extra work for Stripe and can't even use it, but now I guess I always have a fallback if I just want to make it accessible to users without a platform to support it.