Removed home page
It wasn't serving any real purpose, so until I can think of a better use for it I'm going to just not reference the home route anywhere. Now everything that directed to the home page just points to the dashboard. It isn't very pretty but it gets the job done for now.

Also incremented the version counter by 2 because I forgot to do it on the last commit.