Added guide route
Until I can focus enough on the UI for it to be 100% intuitive for all users, this is going to have to do. I explained the basics of the app as simply as I could, but I'm sure I'll end up coming back and tweaking this for a while before anything truly gets fixed.

Ah, the development process.