Moved files to separate dir
Tried replacing jsonwebtoken while I was at it, but I have no way of knowing if it works yet because I can't test it.

I changed all the env references to the right file path so Browserify would work, but I'm going to need to end up replacing it. It isn't secure at ALL.