Application starts
I got it to where the code is running the way it did in the mobile app, with HTML/CSS templating and routing from JavaScript.

OneSignal and Auth0 required new libraries to work on the web instead of Cordova, so I replaced those and modified the CSP to allow them. I'm still working on finding a replacement for the env.js file, so for now I'm including information directly in the code. Only information that doesn't need to be secure, of course, but I'm working on a replacement.