Authentication and dashboard
This is now a working skeleton of the mobile app. You can load data in and authenticate yourself, but other than that I need to change the links around to get any real backend interaction working. I'm about to start working on that first, then once it works I will try to separate UI's for desktop and mobile. Right now the mobile looks just like the app and the desktop looks like shit.