Nicer menu button
Now it will go to the mouse pointer when you're hovering and it's a lot less hacky. I added the jQuery UI library because it makes the menu animation *slightly* better by having the bottom part of the shadow animate with the element and adding the rest in at the end. I don't know why it's being so difficult, but I'm going to stick with this library until I find a better way to do it.