Reauthenticate on login screen
I got rid of the fancy refresh tokens and all that for now. Currently when you are no longer authenticated you are directed back to the login screen. Eventually I'll change that and try to retrieve a new access token first with the callback page, but I'm taking the easiest way that works currently until I get more of the app working.