Web payments and book sharing
Double trouble on this one. I implemented the latest Stripe API to handle payments and I changed some stuff around so you can generate and share books. Still need to touch a lot of things up and really streamline the payments process on my end, but the point is it *sorta* works and I want to track my progress from here, so sue me.