New login flow
I got rid of my nasty login screen that just took you to another Auth0 login screen and just cut out the middleman. Now when you're waiting for the login redirect to finish you get the fancy loading icon, then once you're loaded I changed the colors on Auth0 to match the style of the app.

There was no longer a need for the app login function so I got rid of it. Eventually I want to come back and bring the login screen in-house so I can have custom code in repo and change things a little more to make them prettier, but there are more important tasks to worry about first.