New desktop layout
Not perfect yet, but it's pretty damn good. The mobile layout switches to desktop at 1000px and handles everything semi-elegantly from there. Main issue I need to fix next is with the flexbox centering the last row on the page instead of aligning the items to the rows above, but I need sleep so that's a tomorrow problem.