Using same images across project
The image paths may look a little weird in the context of the repository, but I'll be sure to include a diagram of the web server format in the project README.

The point of this is to drastically speed up loading times, and I have succeeded. Now all the images loaded are loaded from the same location, so the notebook image on the Coming Soon page is already cached from the normal website when you load the page up.

Speaking of, I also brought back the Coming Soon page! This time it's for the web app and not the normal site, even though the normal site isn't looking too great right now. Progress is being made, but it's slow pickings. Anyways, back to it.