Testimonials section
Added in some random filler quotes for now that I need to replace ASAP, but I'm not stressed about it because that's an easy fix down the line. Code isn't the neatest but it gets the job done.

I went ahead and committed some of the About section code because I didn't want to select line by line, but the about section is not ready. I'm going to try working on that in a few days, but today my priority is getting image upload working for the webapp.