Create Book on sidebar
The plus button just wasn't very intuitive, especially on desktop. On mobile this isn't the greatest but I'm going to worry about that just a little bit later. I have a few other ideas for this page but I want to get the backend cemented a little more before making more changes.