New help and redeem page
The redeem code option is now on the sidebar to make it easier to find and to flesh the sidebar out a little more. The guide page was made much more professional with a fancy FAQ dropdown I borrowed from W3. Next step is to try to get a help form up on there so I can respond to user's feedback and stuff.