Help form
The form serializes the data entered and adds it to an SQL database that saves the user's ID, timestamp, auto-increments the primary column, and saves the stringified form data. Next I need to do client-side form validation and add some front end dashboard for me to read the tickets. Eventually I can implement server-side verification as well but that's not priority right now.