Custom ticket dashboard
This is just the start of what I'm planning for the behind-the-scenes stuff for the app. Even though I'm the only one who will ever see it, I want it to be something I'm proud to put on my resume down the line so I can show potential buyers or employers.

Anyways, I can see all the information from the form, I can click their email and immediately have it open in my client for a response (might eventually include a copy of their ticket so they remember what it was about), and whenever I open it for the first time it changes the status from unread to read. Next step is to add flagging, sorting by oldest or newest and different categories, and paging to only retrieve a certain number at a time.