PHP includes for all templates
I wanted to set this up so it would allow me to rely on the same base logic and switch between the actual PHP pages instead of templates to modify the window location, but I couldn't get that to work without a fundamental rework and I don't have time for that right now. So instead I'm settling for making the working environment cleaner and I'm going to have to find a different workaround for the successful payment page and the like.