Removed home route
I'm not going to need that for anything on the web app. I think the dashboard gets the point across, and if I really need to I can go back and add in a tour or something, specifically for mobile, because on desktop everything is easier to use at the moment.