Stats and Notifications
This is probably the most involved commit I've ever made. I decided to add in user stats tracking so gamify it a little with different messages on the sidebar based on your stats, then added the stats to the settings page to flesh it out a little, then noticed that the notifications were broken, so I had to fix those...

Long story short, this fixes most everything with the app. I need to go back in and add stats tracking for a few of the items on the list, then use that info for the mobile menu, but I'll be pretty much set after that. Also gonna spice up the notifications a little if I can.