Cleaning up website
I changed around some of the content because it was stale and unfinished, so hopefully the new stuff is a little better. I fixed up the footer on mobile as well.

I want to go back and add some better fonts before release but I haven't found any that get the job done quite yet, so I'll keep looking.