Back button edge case
Whenever the user clicked on book preview right after creating a new book, then wanted to back out to the create page, it was broken because the handler didn't know what to do. Luckily I created a function for that forever ago and just forgot to implement it in openBookPreview, so this was an easy fix.