New user run-through
There were a few issues that would only arise on the first login then never show again, so I tested for them and was able to resolve them on both the app and webapp. I also optimized the code's layout in a few places to make it easier to follow.