Fireworks on checkout
I wanted to add a little pizzazz to make users feel special, at least as a gimmick of some kind, so the first time they are redirected to the shareCode page after checking out it will have a fullscreen canvas with fireworks going off. Obviously it still needs work, but for now it's good enough and it gets the job done, so I can come back to it if need be down the line.