Improved SEO more
This time I went nearly as far as I could. I minimized and resized every image (in both the images you can see here AND every book and coupon image on the server), added webp formats where they are supported, added better metadata, etc. The only optimizations I was unable to make was CSS loaded at runtime and not asynchronously loading JavaScript, because things get messy when I try to do that. Now we're at the top of the second page of Google and I have my eyes on the first page.