Improved sharing
I was told that when the URL simply said couponbooked.com in the share, people were unable to find the webapp. My next step is to make it easier to navigate to the webapp, but for now I'm just having the webapp's URL in the sharing message.

And although the subject line was helpful for emails, it still showed up in some text messaging applications, resulting in confusion. The safest bet is to remove it for now and users can type in a subject themselves if they want to.