Unbroke things
The www commit was a TERRIBLE idea, because you are never supposed to fix things that aren't broken. Somewhere along the line I broke image uploading on mobile, so I fixed that, then I realized that the pay button issue still exists. So that's the next item on the agenda.